#885 De-complected IEntityDao into reader and mutator contracts.
